OptchainBackend/requirements.txt
Django
djangorestframework
psycopg2-binary
gunicorn
django-cors-headers
yfinance
schedule
holidays